+++ src/lib/tables.h
@@ -1,6 +1,8 @@
 #ifndef TABLES_H
 #define TABLES_H
 
+#include "types.h"
+
 extern const int slen1_tab[16];
 extern const int slen2_tab[16];
 
@@ -8,7 +10,7 @@ extern const long samplerates[9];
 extern const int bitrates[16][4];
 
 extern const int  shine_scale_fact_band_index[9][23];
-extern const long shine_enwindow[];
+extern const int32_t shine_enwindow[];
 
 #endif
